SYMMETRICDS-399 - change the size of the incoming and outgoing batch stat columns to support bigger data loads

+++ trunk/symmetric/symmetric-core/src/main/java/org/jumpmind/symmetric/service/impl/OutgoingBatchService.java
@@ -111,12 +111,12 @@ public class OutgoingBatchService extends AbstractService implements IOutgoingBa
                                 outgoingBatch.getFailedDataId(),
                                 outgoingBatch.getLastUpdatedHostName(),
                                 outgoingBatch.getLastUpdatedTime(), outgoingBatch.getBatchId() },
-                        new int[] { Types.CHAR, Types.INTEGER, Types.INTEGER, Types.INTEGER,
-                                Types.INTEGER, Types.INTEGER, Types.INTEGER, Types.INTEGER,
-                                Types.INTEGER, Types.INTEGER, Types.INTEGER, Types.INTEGER,
-                                Types.INTEGER, Types.INTEGER, Types.INTEGER, Types.INTEGER,
-                                Types.INTEGER, Types.INTEGER, Types.VARCHAR, Types.INTEGER,
-                                Types.VARCHAR, Types.INTEGER, Types.VARCHAR, Types.TIMESTAMP,
+                        new int[] { Types.CHAR, Types.INTEGER, Types.INTEGER, Types.BIGINT,
+                                Types.BIGINT, Types.BIGINT, Types.BIGINT, Types.BIGINT,
+                                Types.BIGINT, Types.BIGINT, Types.BIGINT, Types.BIGINT,
+                                Types.BIGINT, Types.BIGINT, Types.BIGINT, Types.BIGINT,
+                                Types.BIGINT, Types.BIGINT, Types.VARCHAR, Types.INTEGER,
+                                Types.VARCHAR, Types.BIGINT, Types.VARCHAR, Types.TIMESTAMP,
                                 Types.INTEGER });
     }
